fix: change chain_id, chain_length, and chain_pos types from int32_t to int64_t for better range support

+++ src/align/include/align_types.hpp
@@ -28,9 +28,9 @@ namespace align
     float mashmap_estimated_identity;
 
     // Chain metadata
-    int32_t chain_id{-1};               // Unique ID for this chain (-1 if not part of chain)
-    int32_t chain_length{1};            // Total segments in chain (1 if not part of chain)
-    int32_t chain_pos{1};               // Position in chain, 1-based (1 if not part of chain)
+    int64_t chain_id{-1};               // Unique ID for this chain (-1 if not part of chain)
+    int64_t chain_length{1};            // Total segments in chain (1 if not part of chain)
+    int64_t chain_pos{1};               // Position in chain, 1-based (1 if not part of chain)
   };
 
   typedef std::unordered_map <std::string, std::string> refSequenceMap_t;
