+++ src/common/wflign/src/wflign.hpp
@@ -55,9 +55,9 @@ namespace wflign {
             const float min_identity,
             const uint64_t wflign_max_len_minor,
             const float mashmap_estimated_identity,
-            const int32_t chain_id,
-            const int32_t chain_length,
-            const int32_t chain_pos);
+            const int64_t chain_id,
+            const int64_t chain_length,
+            const int64_t chain_pos);
 
         class WFlign {
         public:
